pyproject.toml
[tool.poetry]
name = "fastapi-postgresql-example"
version = "0.1.0"
description = ""
authors = ["Afiz <afiz1337@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
odmantic = { git = "https://github.com/art049/odmantic.git" }
ruff = "^0.0.251"
sqlalchemy = "^2.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
isort = "^5.12.0"
black = "^23"
mypy = "^1.0.1"
pytest-asyncio = "^0.19.0"
pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
